+++ Collatz/upTree/upTree.c
@@ -1,21 +1,21 @@
 #include <assert.h>
 #include "upTree.h"
 struct _upnode {
-    int i;
+    long long i;
     struct _upnode* up;
 };
 
-static struct _upnode* get_head(struct upTree* ut, int key);
-static struct _upnode* get_node(struct upTree* ut, int key);
+static struct _upnode* get_head(struct upTree* ut, long long key);
+static struct _upnode* get_node(struct upTree* ut, long long key);
 static struct _upnode* get_node_head(struct _upnode* node);
 
-struct upTree* ut_create(int array_size) {
+struct upTree* ut_create(size_t array_size) {
     struct upTree* out = malloc(sizeof(struct upTree));
     out->hs = hs_create(array_size);
     return out;
 }
 
-void ut_union(struct upTree* ut, int g1, int g2) {
+void ut_union(struct upTree* ut, long long g1, long long g2) {
     struct _upnode* head1 = get_head(ut, g1);
     struct _upnode* head2 = get_head(ut, g2);
     // I want the group paradigms to be as small as possible
@@ -25,18 +25,18 @@ void ut_union(struct upTree* ut, int g1, int g2) {
         head1->up = head2;
 }
 
-int ut_are_together(struct upTree* ut, int g1, int g2) {
+int ut_are_together(struct upTree* ut, long long g1, long long g2) {
     struct _upnode* head1 = get_head(ut, g1);
     struct _upnode* head2 = get_head(ut, g2);
     return head1 == head2;
 }
 
-static struct _upnode* get_head(struct upTree* ut, int key) {
+static struct _upnode* get_head(struct upTree* ut, long long key) {
     struct _upnode* startNode = get_node(ut, key);
     return get_node_head(startNode);
 }
 
-static struct _upnode* get_node(struct upTree* ut, int key) {
+static struct _upnode* get_node(struct upTree* ut, long long key) {
     struct _upnode** out = (struct _upnode**) hs_get(ut->hs, key);
     if(out == NULL) {
         struct _upnode* new_node = malloc(sizeof(struct _upnode));
@@ -66,7 +66,7 @@ void ut_destroy(struct upTree* ut) {
     free(ut);
 }
 
-int ut_get_group(struct upTree* ut, int g1) {
+long long ut_get_group(struct upTree* ut, long long g1) {
     struct _upnode* n = get_head(ut, g1);
     return n->i;
 }
